refactor: 순환 참조 문제를 해결하기 위해 validate 서비스로 분리
