+++ backend/src/main/java/middle_point_search/backend/domains/memberRoom/MemberRoomService.java
@@ -40,16 +40,4 @@ public class MemberRoomService {
 	public void save(MemberRoom memberRoom) {
 		memberRoomRepository.save(memberRoom);
 	}
-
-	// 회원방 존재 조회
-	public boolean existsByMemberAndRoom(Long memberId, Long roomId) {
-		return memberRoomRepository.existsByMember_IdAndRoom_Id(memberId, roomId);
-	}
-
-	// 방에 존재하는 회원인지 판별
-	public void validateMemberRoom(Long memberId, Long roomId) {
-		if (!memberRoomRepository.existsByMember_IdAndRoom_Id(memberId, roomId)) {
-			throw CustomException.from(MEMBER_ROOM_NOT_FOUND);
-		}
-	}
 }
